reads data sets into tesnor mnist
the data is split into three parts.
1. A training data set, consisting of 55,000 data points. This is used
to adjust weights in the neural network.
2. A validation data set, consisting of 5000 data points. This is used
when training to confirm that you are increasing accuracy with unseen
data points rather than overfitting your network to the training set.
3. A testing data set, consisting of 10,000 data points. This is used to
evaluate the network

Each data point is a 28 by 28 pixel image of digit paired with a value
representing what that digit is (the labels). E.g. mnist.train.images
and mnist.train.labels.

The 28 by 28 image is flattened to a vector of 784 numbers. The
essential quality of flattening is that it be consistent between images.
Once flattened the images can be thought of as data points in a 784
dimensional space (a schema used for
dimensionality reduction).

Therefore mnist.train.images is a tensor of [55000, 784]. The first
dimension is the index of the image; the second is the index of a pixel
in that image of intensity 0 to 1.

Our labels are "one-hot" vectors, (not the parameter "one_hot=True". One
hot vectors mean that all values are 0 except for one value which is 1.
(A "one-cold" vector would have all values at 1 except for one value at
0). The encoding works as follows; the nth digit is a vector with a 1 in
the nth position/dimension. 2 would be [0, 0, 1, 0, 0, 0, 0, 0, 0, 0].
As there are 10 dimensions, the label set mnist.train.labels is an array
of [55000, 10].